Move CMatchConfiguration traits
