+++ include/football/CMatchConfiguration.h
@@ -1,8 +1,7 @@
 #pragma once
 
 #include "IJsonable.h"
-#include "traits/default.h"
-#include "traits/json.h"
+#include "football/traits/CMatchConfiguration.h"
 
 #include "football/CDrawConfiguration.h"
 #include "football/CExtraTime.h"
@@ -13,29 +12,7 @@
 #include "football/CTacticsConfiguration.h"
 #include "football/CTieCondition.h"
 
-namespace futsim
-{
-
-namespace football
-{
-class CMatchConfiguration;
-}
-
-template <> struct default_traits<football::CMatchConfiguration>
-{
-	//! Default \copybrief football::CMatchConfiguration::mApplyAmbientFactor
-	static inline constexpr bool APPLY_AMBIENT_FACTOR = true;
-};
-
-template <> struct json_traits<football::CMatchConfiguration>
-{
-	//! JSON key for the class.
-	static inline constexpr std::string_view KEY = "Match configuration";
-	//! JSON key for the \copybrief football::CMatchConfiguration::mApplyAmbientFactor
-	static inline constexpr std::string_view APPLY_AMBIENT_FACTOR_KEY = "Apply ambient factor";
-};
-
-namespace football
+namespace futsim::football
 {
 
 class CTeamStrategy;
@@ -135,6 +112,4 @@ private:
 	CDrawConfiguration mDrawConfiguration;
 };
 
-} // football namespace
-
-} // futsim namespace
+} // futsim::football namespace
